Project ARES: Nerve Gas Control (#5643)
# About the pull request
Allows nerve gas to be deployed from the AI Core vent system by use of
the APOLLO and ARES Interface consoles.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added Nerve Gas Release control buttons to the ARES and APOLLO
interfaces, allowing the AI core to be secured from large groups or
xenomorphs. Built in cooldowns should prevent fatalities.
/:cl: