Updates UPP Synthetic Loadout (#6098)
# About the pull request

This PR updates the UPP Synthetic loadout to account for it no longer
being a combat synthetic, this includes changing its job title to that
of a Support Synthetic instead of a Combat Synthetic. This effects a
large change to its loadout which is detailed below, and is showcased in
the screenshots section.

Furthermore this PR adds the UPP Support Synthetic job to the UPP crew
monitor, previously niche as this only applied to the survivor UPP
synthetic but now relevant as the regular UPP synthetic is also
considered a support synth.

The major changes to the inventory include.
* Custom UPP synthetic armour, stats identical to USCM Synthetic armour
(no armour, little slowdown)
* Addition of a telescopic baton
* Customized surgical vest and medical storage rig
* Expanded tool pouch
* New epinephrine chem bottle and hypo spray, issued to the UPP synth

# Explain why it's good for the game

The change to the Combat status of the UPP (and CLF) synthetic did not
actually account for their inventory to a complete and comprehensive
degree, ergo these units are lacking in a effective loadout. This
included the fact the synth still technically spawned with armour and
items inside the armour, but the armour (and thus the items inside the
armour) are automatically deleted due to the synth not being able to
wear it, and that the synthetic is still listed in-game as a Combat
Synth.

This PR fixes this and brings the loadout of the UPP synth up to a
better standard to perform its duties as a support Synth whilst giving
it some limited self-defence tools in the baton.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2024-04-06 23 56
38](https://github.com/cmss13-devs/cmss13/assets/6595389/db5ede70-5629-4f84-867b-3a84ad2e1a9f)

</details>


# Changelog
:cl:
balance: UPP Synthetic's loadout has been reworked to account for its
new status as a non-combatant.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>