Project ARES: Orbital Cannon loading log (#6157)
# About the pull request
Adds an ARES log for chambering the OB Cannon, and if it has been
misfueled or not.

# Explain why it's good for the game
Logs who fires it, makes sense to log who loaded it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added an ARES Log for chambering the OB Cannon.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>