Unmarked admin tickets are automatically marked when responding to one (#6167)
# About the pull request
If a ticket is unmarked and you start responding to it, it automatically
marks itself for you.

# Explain why it's good for the game
This is something that's bugged me for a while, and it's always tripped
me up when responding to tickets. One should reason that a ticket you're
responding to is being handled by you.
# Changelog
:cl:
admin: Unmarked tickets now mark themselves when an admin starts
responding to one
/:cl: