Adds an upgraded photocopier for the CL and CC (#6126)
# About the pull request

Adds a new photocopier with a higher toner level (180) and that also
have a higher print-at-once max limit (from the original 10 to 30), the
original photocopier have been maintained for most part of the ship,
receiving a little boost on the toner level (from the original 30 to
45).
Sprites in overall have been updated to have a little bit more charm and
look better, but the new photocopier version have a new "Weyland-Yutani"
paint-scheme (look at the objects' descriptions for reason).

# Explain why it's good for the game

One problem that everyone encounters at least one time playing as CC or
CL is the fact that, the toner limit of the normal photocopier is an
annoying nightmare: Printing photos take a LOT of toner, and even though
that makes a good balancing for the normal photocopier, its a giant ROCK
in the way of anyone who wishes to mass-print newspapers (as CC) or
spread pamphlets and photos (as CL) for the crew. This change would
possibly makes the lives of both roles MUCH easier when printing and
distributing, papers and photos, in mass.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds a new, better variation of the normal photocopier.
imageadd: Adds a new sprite to the new photocopier variation and updates
the sprites of both normal, and new photocopiers to look slightly
better.
maptweak: Adds the new photocopiers to the CC and CL offices.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>