UPP/CLF Combat Synth Presets (#6183)
# About the pull request

Adds a UPP and CLF Combat Synthetic preset, allowing staff to spawn in a
ready to go combat synthetic. These are admin only, they can not spawn
in via regular gameplay.

Their loadouts have been slightly tweaked to retain their current setup
but with an emphasis on having combat weapons.

# Explain why it's good for the game

Adds extra options for staff who may want to allow a Combat Synthetic to
appear in-game for an event, the preset will make it quicker to spawn
one in instead of requiring an admin to fiddle with changing a Synthetic
to a combatant and spawning it in gear.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: UPP and CLF Combat Synthetic preset is now available for admins to
spawn them in.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>