Hugger deletion checking in drone tutorial (#6179)
# About the pull request

This PR is sort of a follow up to #6140 since it became obvious you
could soft-lock yourself in the hugger portion of the tutorial. Rather
than add more checks to prevent huggers deleting in tutorial, I instead
implemented handling when the hugger does die so you can still acquire
another hugger. The 30s deletion timer is still disabled for tutorial
huggers, but jump attempt deaths are possible.

# Explain why it's good for the game

Better replicates the hugger behavior in game, though still not 1-1 and
allows the tutorial to not be soft locked which could be very confusing
for a new player.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/9EBQNe2p4Sw

</details>


# Changelog
:cl: Drathek
fix: Fixed hugger handling in drone tutorial: Now the hugger dying will
put another hugger into the tutorial morpher.
/:cl: