Pill bottle coloring (#6190)
# About the pull request
You can now change pill bottle colors with 12 color options!
Color reference:

![image](https://github.com/cmss13-devs/cmss13/assets/59925169/d8ce2260-adc9-4000-80fe-0106355e9043)

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Allows for some less bland colors on custom pill bottles that are nicer
on the eye
It's also pretty cool
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: You can now change the color of pill bottles in the chem master or
with a hand labeler.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>