Makes spotter laser designators, scout laser designators, and scout cloak unmeltable + indestructible. (#6193)
# About the pull request

The spotter laser designator, scout laser designator and scout cloak are
now unmeltable + indestructible as other irreplaceable specialist gear
is.

# Explain why it's good for the game

The spotter laser designator stood out to me as the most egregious
considering it is mandatory to play the role of spotter effectively - it
should not be destroyable or acidable when it is expected for you to
carry it and when past attempts have been made to enforce reading the
pamphlet and using it. The scout cloak is also OOCly required to be
carried and is irreplaceable + a mandatory part of the role. It seems
more like an oversight than an intended thing to have them meltable.

The scout laser designators I am not entirely certain on making although
it is also irreplaceable and important as a major facet of the scouts
role.


# Testing Photographs and Procedure
I did test this PR to ensure it worked by attempting to melt and firing
a standard OB at the items. They survived.

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
balance: The spotter laser designators, scout laser designators, and
scout cloak are now unmeltable and indestructible to explosions.
/:cl: