Command announcements for survivors (#6192)
# About the pull request
Allows survivors to view tacmap and hear command announcements once they
get IFF and marine headsets.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
When you redeploy as a survivor (which is what you do 90% of the time)
it is extremely painful to play without knowing what's going on, no
announcements, no tacmap no anything. It is also makes sense that you
will have access to those once temporarily enlisting with the marines
and having the proper equipment.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Survivors can now receive command announcements with IFF and a
marine headset.
/:cl: