Adds medlinks to WO and ground maps (#6180)
# About the pull request

This PR is a follow up to #5677 assuming changes from #6103 adding a
medlink to each ground map as well as medlinks for the WO main base.

The other map changes are just the result of StrongDMM's sanitize
variables setting where it deletes var edits that match what the type
already has initially.

# Explain why it's good for the game

Helps establish colony medbays as a temporary objective that can help
the groundside operation with medical supplies. Some maps this is more
practical than others, but the medlinks need to not be too close to a
landing zone else it is too hard to contest them.

As far as WO, there isn't a shipside medical so the main base instead
should serve that purpose.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
maptweak: Added medlinks to WO
maptweak: Added a medlink to each ground map
/:cl: