Mapping LV624 updated apc to the new standart where they autoname and pixel themself to the wall. (#6216)
# About the pull request
created a new subtype of APC for when they have no charge.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: some apc are a bit moved around to avoid them having a small
light on them.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>