Project ARES: Announcement Signatures (#6030)
# About the pull request
ARES now logs who sent an announcement and not just the contents of it.

# Explain why it's good for the game

Makes the available information a little more useful. It's all well and
good recording what was said, but a little pointless if not including
who said it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added the name of an announcement sender, if one exists, to ARES
logs.
/:cl: