Watcher nerf (#6237)
# About the pull request
Watcher facehugger strain now takes damage off weeds the same as the
base variant
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
It's a straight up upgrade to the base variant + Metagaming afk marines
\ open cades from ghost is cancer
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Watcher is now immobilized when zoomed out and loses health off
weeds when not zoomed out
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>