M540-B Armored Recon Carrier (#5401)
# About the pull request
Design doc here: https://hackmd.io/@cmss13/arc

Adds the M540-B Armored Recon Carrier to the game.


![image](https://github.com/cmss13-devs/cmss13/assets/41448081/df012ac6-b21c-4163-8078-f642551dd97a)

The M540 is meant to act as a mobile command & intelligence post for
deploying CIC officers, with all the equipment they'd need (see design
doc for list). It can only be driven by SOs, XOs, COs, and synthetics.


![image](https://github.com/cmss13-devs/cmss13/assets/41448081/6285d198-db71-4157-bc3d-90b461b183af)

The M540 can be obtained for 5 intel points at tier 1. Numbers subject
to change.

The primary features of the ARC are the RE700 rotary cannon sentry and
the U-56 radar antenna. The U-56 takes 8 seconds to retract or deploy.
When deployed, the ARC is anchored in place but has a radar function
akin to the command APC. Since the ARC is now a sitting duck, the RE700
rotary cannon is activated. Any non-USCM personnel or life forms that
come within 5 tiles of the ARC will be fired upon automatically.


![image](https://github.com/cmss13-devs/cmss13/assets/41448081/f3e43a7e-8f5b-4ca1-b6fd-accbe0e21ed0)


# Explain why it's good for the game
Deploying command personnel can be hard-pressed to stay alive or
relevant at times (without just becoming a pointman [COs] or medic+
[SOs]). The ARC gives groundside command personnel the ability to
coordinate and lead from safety, in addition to giving IOs a potential
dropoff point to upload intel and decrypt disks.

# Todo

- [x] Get side-facing movement sprites for the ARC
- [ ] Get a techweb sprite for the ARC

# Changelog
:cl: Zonespace, d.1.n.a. (original sprites), esselnek (new sprites)
add: Added the M540-B Armored Recon Carrier as a t1 intel purchase.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>
Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>