Data detector tweaks (#6194)
# About the pull request

Makes it so that empty vial boxes don't trigger a data detector ping.
Also makes it so that empty randomchem vials don't trigger a data
detector ping.

The former is a bugfix as there is already (non-functional) code in
place to try and achieve this result (see: #3595).
The latter is not a bugfix, but I believe fits nicely into this PR.

*Note that this will still ping vial boxes containing empty vials.
However, I believe this is a fine outcome since there is a low chance
for a vial box to spawn with only empty vials, and I do believe those
should still be pinged until someone checks and empties them.*


# Explain why it's good for the game

There's no reason for empty vial boxes or emptied vials to show up on
the data detector.
Fixing non-functional code is good.


# Testing Photographs and Procedure
<details>
<summary>Testing Evidence</summary>

*Shows that both the intended result of this PR has been achieved and
that nothing else regarding the detector has regressed in the process.*

https://streamable.com/c7hyn6

</details>


# Changelog
:cl:
qol: Emptied randomchem vials no longer show up on the data detector
fix: Fixes empty vial storage boxes showing up on the data detector
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>