Falling Falcons: Paratroopers Edition (#6150)
# About the pull request
Replaces rappel system with paradrop system. You need a parachute on
your back to paradrop (successfully). You also can drop objects from the
above, but in most cases it's not a great idea (they can break). Crates
can be paradropped too, so this can be an interesting alternative to req
drops. In order for marines to paradrop, PO needs to choose a flare on
the console and lock the paradrop system to the signal. After that the
aft hatch will be open and you are free to jump! Just don't forget the
parachute. (SGs can wear parachutes with armor)

Credits to @SubjectD9341 for the sprites.

# Explain why it's good for the game
It's cooler than rappel.

# Testing Photographs and Procedure
<details>


https://github.com/cmss13-devs/cmss13/assets/115417687/53ec9ddb-5ff5-4af2-9df5-17835e5bc2bc


https://github.com/cmss13-devs/cmss13/assets/115417687/f296c1c7-1cfe-4f8f-89c0-233f612d8542


https://github.com/cmss13-devs/cmss13/assets/115417687/579dac1c-9248-4a7c-b0f7-c9466a131335


https://github.com/cmss13-devs/cmss13/assets/115417687/35728d57-7c83-4539-bf84-accb99127ad7

</details>


# Changelog
:cl: ihatethisengine
add: Added paradropping
del: Removed rappeling
/:cl:

---------

Co-authored-by: fira <loyauflorian@gmail.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>