Infection icon now has lower alpha if the human is dead (#6254)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
So this was bothering me for a long time and now I finally found a
simple solution.

I cannot make 5 new sprites to display that hugged person is dead.

I cannot change the color of the sprite, as you need to know which hive
larva belongs to.

But I can lower alpha! Simple solution.

# Explain why it's good for the game
Makes it less confusing for both xenos and observers to figure if the
hugged person is dead. Sometimes xenos ignore people who stepped in a
resin trap because they assume that the person is dead, why would the
person be lying in the middle of nowhere otherwise?

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->

<!-- !! If you are modifying sprites, you **must** include one or more
in-game screenshots or videos of the new sprites. !! -->

![2024-05-07_06-35-46](https://github.com/cmss13-devs/cmss13/assets/115417687/0435546e-f627-4588-a110-387ddfd204ab)

![2024-05-07_06-33-55](https://github.com/cmss13-devs/cmss13/assets/115417687/93ed864a-9b9c-4abf-a64a-532b7c67e686)



# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl: ihatethisengine
add: Infection icon now has lower alpha if the human is dead
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->