Restraints now show up on examine. (#6252)
# About the pull request

As title says, this makes handcuffs/legcuffs appear on examine.

# Explain why it's good for the game

Depending on various overlapping sprites/locations, it can be difficult
to tell if someone is restrained just by looking at them. This adds to
examine something that should have been there to start with.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Handcuffs & Legcuffs now appear on examine.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>