lets you dig out shrapnel with the throwing knife (+ screwdriver) (#6259)
# About the pull request

does what it says on the tin, found this to be mildly irritating that
you couldn't do it

# Explain why it's good for the game

mostly a convenience thing, lets marines eschew the normal bayonet boot
knife for a throwable if they prefer without having to carry the bayonet
in their webbing or rip the one off their gun every time.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
add: you can now remove shrapnel from your body with the M11 Throwing
Knife and screwdrivers
/:cl: