Removes Generic Tailstab for Vamp lurker (#6280)
# About the pull request

Removes vampire lurket tail stab ability as it was intended in the
original design.

# Explain why it's good for the game

Tail stab was not supposed to be on the kit of Vamp lurker . the reason
being is that Tail jab is pretty much the same ( tail stab did not exist
when vampo lurker was designed).
it provides vamp with an insane burst of damage if they movestack both
abilities which no good.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:

balance: Removes Generic tail stab from Vampire lurker
spellcheck: adjusted Vampire lurker description
/:cl: