Adds 70% chance groundside reagent tanks are anchored. (#6279)
# About the pull request

As title.

# Explain why it's good for the game

Whilst Xenos are more than capable of using environmental advantages in
traps, they're not intended to move human equipment to pre-plan traps in
different locations. Rather than rendering it entirely impossible
however, or forbidding it from a rules end, this leaves it to chance.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Added a 70% chance groundside reagent tanks start anchored, if
they don't already.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>