Allows maintenance jack to be quick drawn (#6275)
# About the pull request
Allows maintenance jack to be quick drawn as well as adding a flag that
allows to replicate this behavior for other items.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Quality of life, Requested by a player
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Maintenance jack can now be quick drawn
code: Added a flag that allows items to be quick draw-able despite not
being weapons.
/:cl: