round(x) deprecation for 515 (#6281)
# About the pull request


![image](https://github.com/cmss13-devs/cmss13/assets/14267245/43d08dca-fc72-4401-aabf-ae63460074e7)
> The first format returns the floor of A (the largest integer less than
or equal to A), and has been deprecated in favor of floor(A).

https://www.byond.com/docs/ref/#/proc/round

With the move to 515, replaces any use of the one-argument `round(x)`
with `floor(x)`.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

`floor(x)` clearly communicates what is actually happening, as opposed
to `round(x)` which doesn't.
# Testing Photographs and Procedure
Boots without obvious issue.


# Changelog
No player-facing changes.