Fix MRE automatic trashing and nade box trashing (#6285)
# About the pull request

This PR fixes runtimes with MREs if they spawn on someone that doesn't
have space for them (causing 2 extra qdels than needed), fixes MREs
getting thrown away when you aren't dropping them on the ground, and
changes the behavior of nade_box to not automatically qdel when it is
emptied but instead now has the same behavior as a MRE without food
being dropped on the ground.

# Explain why it's good for the game

Fixes 

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/65dd8860-2728-4a2d-a251-96dfe12878da)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
add: Nade boxes now have the same throw away logic when dropped to the
ground instead of in update_icon when empty.
fix: Fixed MREs getting thrown away if you don't drop them to the
ground.
fix: Fixed MREs throwing runtimes if spawned on a person that doesn't
have space for the MRE.
/:cl: