Changes advanced medhud eye location (#6294)
# About the pull request
For some reason advanced medhud is displayed on a different eye, which
is also extremely ugly looking and has little options to cover up
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The sprites for it look bad and there are few props you can stack ontop
of it to make it look better as opposed to the other eye
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Changed advanced medhud visor to be displayed on the same eye as
normal medhud visor
/:cl: