Maintenance Jack pry time decreased on unpowered doors (#6300)
# About the pull request

This changes the maintenance jack crowbar mode to now open and close
unpowered doors just as fast as a normal crowbar, and allows it to pry
open doors closed. powered doors still take three seconds to pry open.

# Explain why it's good for the game

As it stands, the maintenance jack is a hard sell for synthetics. the
draw of the jack is that it is two tools bundled into one, with benefits
for each mode not found on the individual tools (wrench can open bolts,
crowbar can pry powered doors). The problem with this is that the
crowbar takes forever to open closed and unpowered doors (one of the
most common uses of the crowbar) which makes it practically obsolete
compared to the normal crowbar. Additionally, the jack can't pry doors
closed, a trait exclusive to the jack. This means that not only does it
take forever to pry an unpowered door open, but you also can't pry it
back closed again. This change should lead to more diversity in
synthetic loadouts, as now you aren't using your precious experimental
vendor points on a wrench and a straight up worse crowbar.

# Testing Photographs and Procedure




<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog



:cl: Nomoresolvalou
balance: unpowered doors are now opened instantly in maintenance jack
crowbar mode instead of in three seconds.
balance: crowbar mode can now pry doors closed when unpowered.
/:cl: