Cats now have the correct name and description when held and in your bag (#6306)
# About the pull request

changes the get_scooped proc so that it now uses the correct name and
description when creating a new mob_holder object. This means that mobs
will have the correct name and description when held or in a container.

# Explain why it's good for the game

saying, "Here's the colony cat, Orion!" and then using the show held
item verb, only to show "cat" with the default cat description makes no
sense and breaks my immersion.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog



:cl: Nomoresolvalou

fix: fixed cats not specifically defined in code displaying wrong info
when held or stored in a container. This should also fix any other mob
capable of being picked up that is not defined in code as well.

/:cl: