Roundstart xenomorph count changes (#6291)
# About the pull request
Changes roundstart xeno amounts to try to keep a 1:3 to 1:4 xeno to
combat role ratio
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
https://imgur.com/5dSXlmX
This is just an example but this image is taken pre-drop and there are
23 (25 if you count IOs who do not take part on the front for the most
part) compared to 12 xenos (one died to pred), the goal was always to
keep a 1:3 or 1:4 ratio and it is currently not working and causing
lowpop rounds to be onesided steamrolls.
Every time marines win in lowpop you hear after the round from xenos "we
had 7 burrowed" so yeah something here is wrong.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Latejoin marines to larva ratio changed to 1:4 for the first 15
minutes, going back to 1:2.5 afterwards.
/:cl: